requirements.txt
merlin-dataloader>=23.2
merlin-core>=23.2
pandas>=1.5
scipy>=1.10
pyarrow>=10.0.0
pynvml
setuptools>=62.1.0
numpy>=1.21.0
torch>=1.13.0
functorch>=1.13
lightning
torchmetrics
pytorch-tabnet>=4.0
tensorboard>=2.11
tqdm
matplotlib
scikit-learn
leidenalg
scanpy
rdkit
jupyter
scikit-learn==1.2.1
dgl
dgllife
descriptastorus
torchvision
wandb
plotly
dask_ml
